Refactor BanishInfo - make it a record
