+++ src/main/java/server/life/BanishInfo.java
@@ -1,0 +1,9 @@
+package server.life;
+
+import java.util.Objects;
+
+public record BanishInfo(int map, String portal, String msg) {
+    public BanishInfo {
+        Objects.requireNonNull(portal, "BanishInfo portal");
+    }
+}
