+++ src/main/java/tools/mapletools/MonsterStatFetcher.java
@@ -8,9 +8,9 @@ import provider.DataProviderFactory;
 import provider.DataTool;
 import provider.wz.DataType;
 import provider.wz.WZFiles;
+import server.life.BanishInfo;
 import server.life.Element;
 import server.life.ElementalEffectiveness;
-import server.life.LifeFactory.BanishInfo;
 import server.life.LifeFactory.loseItem;
 import server.life.LifeFactory.selfDestruction;
 import server.life.MobSkillId;
@@ -134,7 +134,10 @@ public class MonsterStatFetcher {
                 }
                 Data banishData = monsterInfoData.getChildByPath("ban");
                 if (banishData != null) {
-                    stats.setBanishInfo(new BanishInfo(DataTool.getString("banMsg", banishData), DataTool.getInt("banMap/0/field", banishData, -1), DataTool.getString("banMap/0/portal", banishData, "sp")));
+                    int map = DataTool.getInt("banMap/0/field", banishData, -1);
+                    String portal = DataTool.getString("banMap/0/portal", banishData, "sp");
+                    String msg = DataTool.getString("banMsg", banishData);
+                    stats.setBanishInfo(new BanishInfo(map, portal, msg));
                 }
 
                 monsterStats.put(mid, stats);
